requirements.txt
# Core
google-genai>=0.2.0
PyGithub>=2.1.1
python-dotenv>=1.0.0

# Utilities
requests>=2.31.0
pydantic>=2.0.0
rich>=13.0.0

# Optional but recommended
tree-sitter>=0.20.0
tree-sitter-python>=0.20.0
gitpython>=3.1.0
# Optional API dependencies
fastapi>=0.104.0
uvicorn>=0.24.0